Add date_prefix configuration option
This change adds an optional new 'date_prefix' configuration option that
will allow for the creation of a date-based prefix in the S3 path. This
will work in combination with the existing 'prefix' option if desired.
Any value for this option is expected to be a time definition suitable
for passing through strftime (e.g '%Y-%m-%d/'). The value of this option
will be inserted between 'prefix' and the filename generated by
logstash. This is particularly useful if you have large volumes of
logs that you wish to keep for long periods of time (archival
purposes). For example, a 5 minute time_file setting will result in
156,960 files for an 18 month retention period for each host sending
data. It is more convenient to work with that number of files,
particularly in GUI based tools, if they are distributed into
smaller groups of files. This will also allow for more granular
loading of files into Elasticsearch.