Delete test files on the remote bucket
We are reverting the changes we did in #85, if we dont delete the files
the files can polute the bucket. So we will try to delete them and
ignore if it fails. Because in reality we only need the permission to
put an object for this plugin to correctly work.

Fixes: #97