Correctly cleanup file when a S3 upload is succesfully
The TemporaryFile#delete! method was raising an exception because we
raryFile#delete! method was raising an exception because we
were trying to close a file that was already close, since the
on_complete is executed in a thread nothing was present to catch and
log that specific error and the thread would just end up getting killed.

This PR add guards for closing an already close file, also add test
around the stale file uploader checker to make sure only 1 file is
remaining on disk after roration.

For future debugging purpuse we also log any error that can occur in the
uploader callback.

Fixes: #122 #120

Fixes #123